Modernise BVH worker & fix XRUI package type config
