+++ __global.d.ts
@@ -61,3 +61,13 @@ declare module '*?worker' {
   const WorkerFactory: { new (): Worker }
   export default WorkerFactory
 }
+
+declare module '*.wasm' {
+  const content: WebAssembly.Module
+  export default content
+}
+
+declare module '*.txt' {
+  const content: string
+  export default content
+}
