+++ packages/client/package.json
@@ -31,7 +31,6 @@
     "react": "18.2.0"
   },
   "dependencies": {
-    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@ir-engine/client-core": "^1.0.3",
     "@ir-engine/common": "^1.0.3",
     "@ir-engine/editor": "^1.0.3",
@@ -42,6 +41,7 @@
     "@originjs/vite-plugin-commonjs": "^1.0.3",
     "app-root-path": "3.1.0",
     "cli": "1.0.1",
+    "esbuild-plugins-node-modules-polyfill": "^1.7.1",
     "history": "^5.3.0",
     "i18next": "21.6.16",
     "i18next-browser-languagedetector": "6.1.3",
