+++ packages/spatial/src/transform/functions/generateBVHAsync.worker.js
@@ -1,29 +1,39 @@
-
-
-
-// TODO: Avoid code duplication: There is a copy of this script in /public
-
-import * as THREE from 'three'
+import { BufferAttribute, BufferGeometry } from 'three'
 import { MeshBVH } from 'three-mesh-bvh'
 
 addEventListener('message', ({ data }) => {
-  const { index, position, options } = data
+  const { index, position, groups, options } = data
 
   try {
-    const geometry = new THREE.BufferGeometry()
-    geometry.setAttribute('position', new THREE.BufferAttribute(position, 3))
+    const geometry = new BufferGeometry()
+    geometry.setAttribute('position', new BufferAttribute(position, 3, false))
     if (index) {
-      geometry.setIndex(new THREE.BufferAttribute(index, 1))
+      geometry.index = new BufferAttribute(index, 1, false)
+    }
+    if (groups) {
+      for (const group of groups) {
+        geometry.addGroup(group.start, group.count, group.materialIndex)
+      }
     }
     options.lazyGeneration = false
-    const bvh = new MeshBVH(geometry, options)
-    const serialized = MeshBVH.serialize(bvh, { copyIndexBuffer: false })
+    const bvh = new MeshBVH(geometry, { setBoundingBox: false, ...options })
+    const serialized = MeshBVH.serialize(bvh, { cloneBuffers: false })
+
+    let transferrables = [position.buffer, ...serialized.roots]
+    if (serialized.index) {
+      transferrables.push(serialized.index.buffer)
+    }
+    if (bvh._indirectBuffer) {
+      transferrables.push(serialized.indirectBuffer.buffer)
+    }
+
     postMessage(
       {
         error: null,
-        serialized
+        serialized,
+        groups: geometry.groups
       },
-      [serialized.index.buffer]
+      transferrables
     )
   } catch (error) {
     postMessage({
