+++ packages/xrui/core/textures/KTX2Worker.ts
@@ -1,5 +1,4 @@
 import BasisEncoderModuleSRC from './basis_encoder_low_memory/basis_encoder.js.txt'
-// @ts-ignore
 import BasisEncoderWASMBinary from './basis_encoder_low_memory/basis_encoder.wasm'
 import type { KTX2EncodeRequestData, KTX2EncodeResponseData } from './KTX2Encoder'
 ;(0, eval)(BasisEncoderModuleSRC)
